service_a/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
PyJWT==2.11.0
argon2-cffi==25.1.0
pika==1.3.2
python-dotenv==1.2.2
pytest==9.0.2
httpx==0.28.1
pydantic-settings==2.13.1
email-validator==2.3.0
python-multipart==0.0.22
